// src/enum.ts
export enum AppRouteEnum {
    Home = '/',
    Blog = '/blog',
    About = '/about',
    NotFound = '/404',
    Podcast = '/podcast',
    Work = '/portfolio/work',
    Portfolio = '/portfolio',
    Study = '/portfolio/study',
    Devotion = '/portfolio/devotion',
    Accumulation = '/portfolio/accumulation',
}